Move utils into own classes
